Uploaded local work to github
This is the project as it stands on 6/28. There is no system for cleanly inputting terrain, and minimal bugtesting has been done. Much of the component/weapon data is incomplete, especially at higher tech levels. All of the game's mechanics and features are coded however, and (hopefully) are working properly.